major update! able to run test cases for Two Sum, improved result display UI.

+++ src/main/java/org/johnnyhe/dayleet/model/CodeSubmissionRequest.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class CodeSubmissionRequest {
-    private Long questionId;
+    private int questionId;
     private String code;
     private Integer languageId;
